refactor: move yaml system into package
