+++ src/libecalc/input/yaml_types/components/system/yaml_compressor_system.py
