+++ src/libecalc/input/yaml_types/components/yaml_generator_set.py
@@ -5,11 +5,11 @@ from libecalc.input.yaml_types import YamlBase
 from libecalc.input.yaml_types.components.legacy.yaml_electricity_consumer import (
     YamlElectricityConsumer,
 )
-from libecalc.input.yaml_types.components.yaml_category_field import CategoryField
-from libecalc.input.yaml_types.components.yaml_compressor_system import (
+from libecalc.input.yaml_types.components.system.yaml_compressor_system import (
     YamlCompressorSystem,
 )
-from libecalc.input.yaml_types.components.yaml_pump_system import YamlPumpSystem
+from libecalc.input.yaml_types.components.system.yaml_pump_system import YamlPumpSystem
+from libecalc.input.yaml_types.components.yaml_category_field import CategoryField
 from libecalc.input.yaml_types.yaml_placeholder_type import YamlPlaceholderType
 from libecalc.input.yaml_types.yaml_schema_helpers import (
     replace_temporal_placeholder_property_with_legacy_ref,
